conftest: move complex_codeplug to shared
